+++ doc/changelog.rst
@@ -6,6 +6,9 @@ unreleased: Version 0.14.0
 
 New features:
 
+- add type annotations and make types stricter PR # 105
+  In particular, ConcreteInstr does not inherit from Instr anymore and one
+  cannot use ConcreteInstr in Bytecode object. This is saner than before.
 
 Bugfixes: